+++ Weitianshi/app.json
@@ -2,7 +2,6 @@
     "pages": [
       
         "pages/match/match/match/match",
-       "pages/oneKeyInfo/oneKeyInfo",
         "pages/my/my/my",
         "pages/search/filter4/filter4",
         "pages/match/match/resourceDemand/resourceDemand",
@@ -11,11 +10,10 @@
         "pages/form/industry/industry",
         "pages/form/scale/scale",
         "pages/form/stage/stage",
-        "pages/projectDetai/projectDetai",
 "pages/form/businessContribution/businessContribution",
         "pages/form/tollType/tollType",
         "pages/form/hasRunningBusiness/hasRunningBusiness",
-        "pages/scanCode/bpScan/bpScan",
+      "pages/scanCode/bpScan/bpScan",
         "pages/scanCode/bpScanSuccess/bpScanSuccess",
         "pages/scanCode/cardScan/cardScan",
         "pages/scanCode/cardScanEdit/cardScanEdit",
@@ -28,6 +26,7 @@
         "pages/identity/investorFA/investorFA",
         "pages/identity/entrepreneurFA/entrepreneurFA",
         "pages/message/message/message",
+        "pages/message/viewProjectUser/viewProjectUser",
         "pages/message/applyMyProject/applyMyProject",
         "pages/message/myApplyProject/myApplyProject",
         "pages/message/bePushedProject/bePushedProject",
