+++ Weitianshi/pages/message/viewProjectUser/viewProjectUser.json
@@ -1,0 +1,3 @@
+{
+  "navigationBarTitleText": "浏览项目的人"
+}
