+++ Weitianshi/pages/message/viewProjectUser/viewProjectUser.wxss
@@ -1,0 +1,326 @@
+/* pages/message/viewProjectUser/viewProjectUser.wxss */
+.browseMeTitle {
+  /*height: 90rpx;*/
+  margin: 32rpx 32rpx 32rpx 32rpx;
+}
+
+.flex {
+  display: flex;
+}
+
+.title {
+  font-size: 26rpx;
+  color: #333;
+  line-height: 26rpx;
+  text-align: left;
+}
+
+.nothing {
+  font-size: 36rpx;
+  width: 146rpx;
+  height: 50rpx;
+  color: #999;
+  text-align: left;
+  margin: 65% auto;
+}
+
+.title .block {
+  background: #40587a;
+  width: 4rpx;
+  height: 24rpx;
+  margin-right: 12rpx;
+}
+
+.redPoint {
+  background: #f44c4c;
+  width: 16rpx;
+  height: 16rpx;
+  border-radius: 100%;
+}
+
+.browseMeTitle .num {
+  font-size: 24rpx;
+  color: #666;
+  line-height: 24rpx;
+  text-align: right;
+}
+
+/*--------------------------*/
+
+page {
+  height: 100%;
+}
+
+.s_frame {
+  height: 1052rpx;
+  position: static;
+}
+
+.nullState_text {
+  text-align: center;
+}
+
+.nullState_text image {
+  width: 176rpx;
+  height: 176rpx;
+  margin-top: 248rpx;
+}
+
+.nullState_text view:nth-child(2) {
+  font-size: 32rpx;
+  padding-top: 40rpx;
+  padding-bottom: 24rpx;
+  font-weight: 600;
+}
+
+.nullState_text view:nth-child(3) {
+  color: #999;
+  font-size: 28rpx;
+  padding-bottom: 60rpx;
+}
+
+.nullState button {
+  width: 360rpx;
+  margin: 0 auto;
+  height: 88rpx;
+  line-height: 88rpx;
+  font-size: 30rpx;
+  color: white;
+  background: #40587a;
+  border-radius: 8rpx;
+  padding: 0;
+}
+
+/*用户信息*/
+
+.card {
+  background: white;
+  width: 100%;
+  padding: 32rpx 38rpx 0 0;
+}
+
+.user_info {
+  margin-bottom: 20rpx;
+  margin-left: 32rpx;
+}
+
+.user_text, .user_pic {
+  width: 120rpx;
+  height: 120rpx;
+  border-radius: 50%;
+}
+
+.user_text {
+  background: #f2f2f2;
+  line-height: 120rpx;
+  text-align: center;
+  font-size: 48rpx;
+}
+
+.user_info_right {
+  margin-left: 24rpx;
+}
+
+.frame {
+  justify-content: space-between;
+  width: 540rpx;
+}
+
+.name_flex {
+  justify-content: space-between;
+  align-items: center;
+}
+
+.user_name {
+  font-size: 32rpx;
+  color: #333;
+}
+
+.name_flex .user_career {
+  font-size: 24rpx;
+  color: #666;
+  padding-top: 16rpx;
+  margin-left: 16rpx;
+}
+
+.user_company {
+  font-size: 24rpx;
+  color: #666;
+  padding-top: 24rpx;
+}
+
+.pro_time {
+  background: #f2f2f2;
+  border-radius: 8rpx;
+  width: 86rpx;
+  height: 28rpx;
+  line-height: 28rpx;
+  color: #999;
+  font-size: 20rpx;
+  margin-left: 272rpx;
+}
+
+.swiper-box .user_company {
+  text-align: left;
+  padding-top: 12rpx;
+  font-size: 20rpx;
+  color: #999;
+}
+
+.pro_info {
+  margin-top: 12rpx;
+  width: 622rpx;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.swiper-box .pro_intro {
+  padding: 20rpx 0;
+  font-size: 28rpx;
+  color: #333;
+  text-align: left;
+}
+
+.industry, .stage {
+  font-size: 24rpx;
+  color: #666;
+  padding: 0 16rpx;
+}
+
+.industry {
+  border-right: 1px solid #e0e0e0;
+}
+
+.industry:first-child {
+  padding-left: 0;
+}
+
+.industry:last-child {
+  border: 0;
+}
+
+.stage:first-child {
+  padding-left: 0;
+}
+
+.investor_right {
+  flex-wrap: wrap;
+  width: 580rpx;
+}
+
+.industry_tip {
+  font-size: 24rpx;
+  color: #666;
+  padding: 0 16rpx;
+  border-right: 1px solid #e0e0e0;
+  margin-bottom: 20rpx;
+}
+
+.industry_tip:first-child {
+  padding-left: 0;
+}
+
+.industry_tip:last-child {
+  border: none;
+}
+
+.pro_sort {
+  margin-right: 24rpx;
+  font-size: 20rpx;
+  color: #999;
+  border-radius: 98rpx;
+  border: 1px solid #e0e0e0;
+  padding: 0 16rpx;
+  height: 26rpx;
+  line-height: 26rpx;
+}
+
+/*投资需求*/
+
+.investor_intro, .investor_industry_frame, .investor_stage_frame {
+  align-items: baseline;
+}
+
+.investor_text {
+  font-size: 24rpx;
+  flex-wrap: nowrap;
+  width: 126rpx;
+  margin: 12rpx 0;
+  text-align: left;
+}
+
+.investor_text, investor_right {
+  padding: 0 12rpx;
+}
+
+/*资源需求*/
+
+.res_give, .res_find {
+  font-size: 24rpx;
+  border: 1px solid #ccc;
+  color: #ccc;
+  padding: 0 12rpx;
+  border-radius: 8rpx;
+  height: 46rpx;
+  line-height: 46rpx;
+  margin-right: 24rpx;
+  margin-bottom: 16rpx;
+}
+
+.addNetwork {
+  background: #40587a;
+  border-radius: 200rpx;
+  width: 120rpx;
+  height: 48rpx;
+  color: #fff;
+}
+
+.addNetwork .addNet {
+  font-size: 20rpx;
+  color: #fff;
+  line-height: 54rpx;
+  display: block;
+  text-align: center;
+}
+
+.netWork {
+  opacity: 0.8;
+  font-family: PingFangSC-Regular;
+  font-size: 28rpx;
+  color: #999;
+  letter-spacing: 0;
+  line-height: 28rpx;
+  text-align: left;
+}
+
+.addPerson {
+  border-radius: 200rpx;
+  width: 120rpx;
+  height: 48rpx;
+  position: static;
+}
+
+.addPersonBtn {
+  background: #40587a;
+  font-family: PingFangSC-Regular;
+  font-size: 20rpx;
+  border-radius: 100px;
+  color: #fff;
+  padding: 14rpx 0;
+  line-height: 20rpx;
+  text-align: center;
+  position: static;
+}
+
+.addPersonBtn::after {
+  border: 0;
+}
+
+.addTime {
+  font-size: 20rpx;
+  padding-top: 10rpx;
+  padding-right: 32rpx;
+  padding-bottom: 20rpx;
+  text-align: right;
+  color: #999999;
+}
