--- Weitianshi/pages/oneKeyInfo/oneKeyInfo.json
@@ -1,3 +1,0 @@
-{
-    "navigationBarTitleText": "一键尽调"
-}
