--- Weitianshi/pages/projectDetai/projectDetai.json
@@ -1,1 +1,0 @@
-{"navigationBarTitleText": "项目管理"}
