--- Weitianshi/pages/projectDetai/projectDetai.wxml
@@ -1,46 +1,0 @@
-<!--pages/projectDetai/projectDetai.wxml-->
-
-<view class="classname">
-  <view class="projectDetailTop">
-    <view class="projectDetailBox">
-      <view class="flex">
-        <view class="describle">
-          一句话描述项目亮点,大概应该是40个字左右,要全部显示
-        </view>
-        <view class="projectDetailType">非獨家</view>
-      </view>
-      <view class="industry">人工智能、大数据、理财、企业服务台</view>
-      <view class="allProjectDetail flex">
-        <view class="classname">300-500w万</view>
-        <view class="classname">10%</view>
-        <view class="classname">天使轮</view>
-        <view class="classname">杭州</view>
-      </view>
-      <view class="ProjectClues flex">
-        <view class="classname">项目线索</view>
-        <view class="ProjectCluesLine"></view>
-        <view class="classname">佣金收讫</view>
-      </view>
-      <view class="ProjectDetailInfo flex">
-        <view class="sceenPerson">浏览：1000</view>
-        <view class="applyPerson">申请查看：-</view>
-        <view class="investPerson">意向投资人：-</view>
-      </view>
-    </view>
-  </view>
-
-  <swiper current="{{currentTab}}" class="swiper-box" duration="300" style="height:{{winHeight - 31}}px" bindchange="bindChange">  
-    <!-- 我是哈哈 -->  
-    <swiper-item>  
-      <view>我是哈哈</view>  
-    </swiper-item>  
-    <!-- 我是呵呵 -->  
-    <swiper-item>  
-      <view>我是呵呵</view>  
-    </swiper-item>  
-    <!-- 我是嘿嘿 -->  
-    <swiper-item>  
-      <view>我是嘿嘿</view>  
-    </swiper-item>  
-</swiper>  
-</view>
