+++ Weitianshi/pages/projectDetail/projectDetail.js
@@ -56,11 +56,12 @@ Page({
                   success: function (res) {
                     console.log(res)
                     var button_type = res.data.button_type;
+                    let userInfo = res.data.user_info;
                     console.log(user_id)
                     console.log(button_type)
                     that.setData({
                       button_type: button_type,
-
+                      userInfo: userInfo
                     })
                   }
                 })
