+++ Weitianshi/pages/projectDetail/projectDetail.wxml
@@ -1,80 +1,63 @@
 <scroll-view scroll-y="true">
-
   <!--头部名片-->
   <view id="myDetail">
-    <view class="weui-panel__bd" style="background:white;">
-
+    <view class="weui-panel__bd">
       <!--名片-->
-      <view class="weui-cell" style="background:white;margin:32rpx;margin-bottom:0;border-radius:4px 4px 0 0" bindtap="userDetail" data-id="{{user.user_id}}">
-        <view class="weui-cell__hd" style="position: relative;margin-right: 10px;">
-          <image wx:if="{{user.user_avatar_url}}" class="others_pic" src="{{user.user_avatar_url}}" />
-          <view wx:if="{{!user.user_avatar_url}}" class="others_nopic">{{user.user_avatar_text}}</view>
-        </view>
-        <view class="weui-cell__bd">
-          <view class="flex" style="align-items: baseline;">
-            <view class="others_name">{{user.user_name}}</view>
-            <view class="others_email">{{user.user_email}}</view>
-          </view>
-          <view class="flex">
+      <view class="classname">
+        <view class="projectDetailTop">
+          <view class="projectDetailBox" bindtap="toMy">
+            <view class="flex">
+              <view class="describle">{{project.pro_intro}}</view>
+              <view class="theOne exclusive" wx:if="{{project.is_exclusive==1}}">独家签约</view>
+              <view class="theOne exclusive" wx:if="{{project.is_exclusive==2}}">非独家签约</view>
+            </view>
+            <view class="industry" wx:key="{{item.industry_name}}">人工智能、大数据、理财、企业服务台</view>
+            <view class="allProjectDetail flex">
+              <view class="classname">{{project.pro_scale.scale_money}}</view>
+              <view class="classname">10%</view>
+              <view class="classname">{{project.pro_stage.stage_name}}</view>
+              <view class="classname">{{project.pro_area.area_title}}</view>
+            </view>
+            <view class="otherInfo flex" bindtap="userDetail" data-id="{{user.user_id}}">
+              <view class="weui-cell__hd">
+                <image wx:if="{{user.user_avatar_url}}" class="others_pic" src="{{user.user_avatar_url}}" />
+                <view wx:if="{{!user.user_avatar_url}}" class="others_nopic">{{user.user_avatar_text}}</view>
+              </view>
+              <view class="weui-cell__bd">
+                <view class="flex userInfo" style="align-items: baseline;">
+                  <view class="userName">{{user.user_name}}</view>
+                  <view class="userCareer">{{userInfo.user_company_career}}</view>
+                  <view class="userCompany">{{userInfo.user_company_name}}</view>
+                  <image src="/img/icon-gengduo2@2x.png" class="userIcon"></image>
+                </view>
+                <!--<view class="flex">
             <view class="others_company">{{user.user_company}} | {{user.user_career}}</view>
             <view class="addPerson" wx:if="{{button_type == 0 ||button_type == 3||button_type == 4}}">
-              <view class="addPersonBtn" catchtap="addPerson" data-id="{{user.user_id}}" >加人脉</view>
+              <view class="addPersonBtn" catchtap="addPerson" data-id="{{user.user_id}}">加人脉</view>
             </view>
-             <view class="waitPerson" wx:if="{{button_type == 1 && followed_user_id!==user_id}}">
+            <view class="waitPerson" wx:if="{{button_type == 1 && followed_user_id!==user_id}}">
               <view class="waitPersonBtn" style="background:#ffffff;color:#cccccc;">已为人脉</view>
             </view>
             <view class="waitPerson" wx:if="{{button_type == 2}}">
               <view class="waitPersonBtn" style="background:#ffffff;color:#cccccc;">待验证</view>
             </view>
-          </view>
-          <!--<view class="ohters_position"></view>-->
-        </view>
-      </view>
-      <view class="greyFrame">
-
-        <!--项目完成度-->
-
-        <!--<view class="complete">项目完整度{{complete}},打败了{{loser}}的项目</view>-->
-
-        <!--项目简介-->
-        <view class="weui-media-box weui-media-box_text">
-          <view class=" box__title_in-text">
-            {{project.pro_intro}}
-            <view class="theOne exclusive" wx:if="{{project.is_exclusive==1}}">独家签约</view>
-            <view class="theOne exclusive" wx:if="{{project.is_exclusive==2}}">非独家签约</view>
-          </view>
-
-          <!--标签1-->
-          <view class=" sort ">
-            <block wx:for="{{project.pro_industry}}" wx:key="{{item.industry_name }}">
-              <view class="weui-media-box__desc " wx:if="{{index<4}}">
-                {{item.industry_name}}
+          </view>-->
+                <!--<view class="ohters_position"></view>-->
               </view>
-            </block>
-          </view>
-
-          <!--标签2-->
-          <view class="tips flex ">
-            <view class="flex ">
-              <view style="width:1px"></view>
-              <view class="weui-media-box__desc tip " wx:if="{{project.pro_stage.stage_name}}">{{project.pro_stage.stage_name}}</view>
-              <view class="weui-media-box__desc tip " wx:if="{{project.pro_area.area_title}}">{{project.pro_area.area_title}}</view>
-              <view class="weui-media-box__desc tip " wx:if="{{project.pro_scale.scale_money}}">{{project.pro_scale.scale_money}}</view>
             </view>
-            <view class="tip flex ipo" wx:if="{{stock!==0}}">投后股份:
-              <view style="color:red ">{{stock}}%</view>
-            </view>
-          </view>
-
-          <!--项目亮点-->
-          <view class="greatPoint" wx:if="{{project.pro_goodness!==''}}">
-            <view>项目亮点</view>
-            <view>{{project.pro_goodness}}</view>
           </view>
         </view>
-        <view class='ziben'>—— 项目详情 ——</view>
       </view>
-
+      <!--项目亮点-->
+      <view class="brightBox" style="background: white;">
+       <view class="brightPoint">
+            <view class="brightPointBox flex">
+              <image src="/img/img-star.png" class="titlePic"></image>
+              <view class="title star">项目亮点</view>
+            </view>
+            <view class="brightContent" bindtap="more" value="{{number}}">{{project.pro_goodness}}</view>
+          </view>
+          </view>
       <!--项目详情-->
       <view class="detail">
         <view class="grey" wx:if="{{bpName!==''}}">项目BP</view>
@@ -96,7 +79,6 @@
       <image class="lock" src="/img/lock.png"></image>
 
       <!--资本对接-->
-
       <!--<view class=" resource ">-->
       <view class="others">
         <block wx:for="{{investor}}" wx:key="{{investor_id}}" wx:for-item="{{investor}}">
@@ -108,10 +90,8 @@
             <view class="weui-cell__bd ">
               <view class="flex " style="align-items: baseline; ">
                 <view class="others_name ">{{item.investor_name}}</view>
-
                 <!--<view class="others_email ">786485758@qq.com</view>-->
               </view>
-
               <!--<view class="others_company" wx:if="{{item.investor_type==2}}">{{item.investor_company}}</view>-->
               <view class="ohters_position" wx:if="{{item.investor_type==2}}">{{item.investor_career}}</view>
             </view>
