+++ Weitianshi/pages/projectDetail/projectDetail.wxss
@@ -1,5 +1,5 @@
 page {
-  background: white;
+  /*background: white;*/
   height: 100%;
 }
 
@@ -82,6 +82,8 @@ scroll-view {
 }
 
 .theOne {
+  position: absolute;
+  right: 48rpx;
   background: #40587a;
   color: white;
   display: inline-block;
@@ -136,17 +138,18 @@ scroll-view {
   color: #666;
 }
 
-.others_company {
+/*.others_company {
   font-size: 24rpx;
   color: #666;
   padding-bottom: 16rpx;
-}
+}*/
 
+/*
 .others_email {
   font-size: 24rpx;
   color: #333;
   padding-left: 16rpx;
-}
+}*/
 
 .others_tips {
   border: 1px solid #333;
@@ -227,7 +230,7 @@ scroll-view {
 .floatBtn button {
   width: 100%;
   justify-content: center;
-  color: #ccc;
+  color:#ffffff;
   border-radius: 0;
   background: #f2f2f2;
 }
@@ -270,19 +273,26 @@ scroll-view {
 /*头像*/
 
 .others_pic {
-  width: 120rpx;
-  height: 120rpx;
+  width: 48rpx;
+  height: 48rpx;
   display: inline-block;
   text-align: center;
   border-radius: 50%;
+  margin-top: 18rpx;
+  margin-right: 16rpx;
 }
 
 .others_nopic {
-  font-size: 64rpx;
+  font-size: 32rpx;
   color: #40587a;
   background: #f2f2f2;
-  padding: 28rpx 40rpx;
+  padding: 14rpx 20rpx;
   border-radius: 50%;
+  width: 48rpx;
+  height: 48rpx;
+  text-align: center;
+  margin-top: 18rpx;
+  margin-right: 16rpx;
 }
 
 .load {
@@ -294,7 +304,7 @@ scroll-view {
   height: 160rpx;
   padding-left: 295rpx;
   padding-top: 170rpx;
-  padding-bottom: 295rpx;
+  /*padding-bottom: 295rpx;*/
 }
 
 /*加人脉*/
@@ -362,8 +372,8 @@ scroll-view {
 }
 
 .waitPersonBtn {
-  background:#ffffff;
-  border:1rpx solid #cccccc;
+  background: #fff;
+  border: 1rpx solid #ccc;
   border-radius: 100px;
   width: 118rpx;
   /*height: 48rpx;*/
@@ -372,4 +382,167 @@ scroll-view {
   line-height: 20rpx;
   text-align: center;
   position: static;
+}
+
+/*新的内容*/
+
+.projectDetailTop {
+  /*height: 420rpx;*/
+  width: 90%;
+  height: auto;
+  background: #fff;
+  box-shadow: 0 8rpx 8rpx 0 rgba(64, 88, 122, 0.10);
+  border-radius: 8rpx;
+  margin: 32rpx 32rpx 38rpx 36rpx;
+}
+
+.projectDetailBox {
+  padding: 32rpx;
+  height: auto;
+  padding-bottom: 0;
+}
+
+.describle {
+  font-size: 32rpx;
+  color: #333;
+  line-height: 48rpx;
+  text-align: left;
+}
+
+.projectDetailType {
+  border: 2rpx solid #40587a;
+  border-left: none;
+  width: 104rpx;
+  height: 32rpx;
+  font-size: 20rpx;
+  color: #40587a;
+  line-height: 32rpx;
+  text-align: right;
+}
+
+.industry {
+  font-family: PingFangSC-Regular;
+  font-size: 26rpx;
+  color: #666;
+  line-height: 26rpx;
+  text-align: left;
+  margin: 36rpx 0 24rpx 0;
+}
+
+.allProjectDetail {
+  font-size: 26rpx;
+  color: #666;
+  line-height: 26rpx;
+  text-align: left;
+  margin-bottom: 34rpx;
+}
+
+.ProjectCluesLine {
+  width: 67%;
+  height: 2rpx;
+  margin: 10rpx 18rpx 0 18rpx;
+  font-size: 0;
+  background: #e6e6e6;
+}
+
+.otherInfo {
+  border-top: 1rpx solid #e6e6e6;
+  height: 94rpx;
+}
+
+.userCareer {
+  font-size: 24rpx;
+  color: #666;
+  letter-spacing: 0;
+  line-height: 24rpx;
+  text-align: left;
+  margin-right: 16rpx;
+  margin-top: 34rpx;
+}
+
+.userCompany {
+  font-size: 20rpx;
+  color: #666;
+  letter-spacing: 0;
+  line-height: 20rpx;
+  text-align: left;
+  margin-top: 36rpx;
+}
+
+.userName {
+  font-size: 28rpx;
+  color: #3a3e4a;
+  letter-spacing: 0;
+  line-height: 28rpx;
+  text-align: left;
+  margin-right: 16rpx;
+  margin-top: 32rpx;
+}
+
+.userIcon {
+  position: absolute;
+  top: 40rpx;
+  right: 2rpx;
+  width: 12rpx;
+  height: 16rpx;
+}
+
+.userInfo {
+  position: relative;
+}
+
+/*项目亮点*/
+
+.brightBox {
+  padding: 32rpx;
+  text-align: left;
+}
+
+.brightPoint {
+  background: #fafafa;
+  border-radius: 2px;
+  padding: 32rpx 24rpx 32rpx 24rpx;
+}
+
+.titlePic {
+  width: 28rpx;
+  height: 28rpx;
+  padding-right: 14rpx;
+  display: inline-block;
+}
+
+.star {
+  margin-bottom: 26rpx;
+}
+
+.brightContent {
+  height: auto;
+  font-size: 26rpx;
+  color: #666;
+  line-height: 39rpx;
+  text-align: left;
+  text-overflow: -o-ellipsis-lastline;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  background: #fafafa;
+}
+
+.title {
+  font-size: 28rpx;
+  color: #333;
+  line-height: 28rpx;
+  text-align: left;
+}
+
+.push {
+  height: 88rpx;
+  font-size: 28rpx;
+  line-height: 88rpx;
+}
+
+button::after{
+  border: none;
 }
