+++ Weitianshi/pages/scanCode/bpScan/bpScan.json
@@ -1,3 +1,3 @@
 {
-    "navigationBarTitleText": "用户详情"
+    "navigationBarTitleText": "项目详情"
 }
